accounts: add force: bool arg to load()
Sometimes watch futures from backends send refresh events for mailboxes
that have not been fetched yet. That'd trigger fetching them in the
background, which is bad when the mailboxes are big and not wanted by
the user in the first place.

This commit adds a force parameter that specifies whether the loading
must be forced (e.g. when wanting to display the mailbox) or not.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>