CI: .gitea/Makefile.lint: check if nightly exists
If nightly toolchain exists and `cargo +nightly fmt --check --all` runs
but fails because --check returns an error, and the same error does not
trigger under the current active toolchain when the `|| cargo fmt
--check --all` path is executed, then that error is completely ignored.

That wouldn't happen on the CI where the nightly toolchain is not
installed but it would happen if someone (i.e. me) used the Makefiles
locally.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>