melib/backend: remove BackendOp
This trait used to hold more methods than just returning an e-mail as
bytes. But that's all it did nowadays. Remove it since it's not really
necessary to have a level of indirection just to fetch an envelope as
bytes from a mail backend.

The `MailBackend::operation` method is replaced by a
`MailBackend::envelope_bytes_by_hash` method which does the same thing
as calling `MailBackend::operation` and then await on the `as_bytes`
method of its return value.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>