melib/error: hardcode native-tls error check
When converting an std::io::Error into an Error, check if it's an error
returned by the platform implementation of native-tls by looking for the
strings "the SSL session has been shut down" or "OpenSSL error".

This way, we can convert it into an ErrorKind::Network and the frontend
can detect that it's safe to retry a connection when this error is
returned.

Note: this error was observed in the NNTP backend, probably because the
keep-alive logic for the NNTP connection is different than IMAP.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>