melib/imap: retry IDLE with exponential backoff
When an IDLE connection starts, it selects the INBOX and issues the
`IDLE` IMAP command. However, if the main connection hasn't finished
retrieving the mailboxes yet, INBOX might not be available and the
watch connection will error out, and the front-end will restart it.

Instead of failing right away, retry at a random time interval at most
10 times before failing.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>