mail/listing: restore selection when refreshing same mailbox
If you start making a selection and `refresh_mailbox` is called, maybe
because the mailbox is still loading or because it was manually
refreshed, the selection is lost.

This commit changes the `RowsState::clear` method to optionally return
the previous selection and then restore it with a new
`RowsState::restore_selection` method.

Fixes #569 ("Problem with visual mode")

Resolves: https://git.meli-email.org/meli/meli/issues/569
Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>