melib/mbox: require MboxFormat for parsing
It seems that when no format is specified and the default fallback
mboxcl2 is selected and the mbox file contains emails that contain
invalid Content-Length values, it's possible for the parsing logic to
lose track of the correct byte offsets because it trusts the
Content-Length it reads.

This also happens in non-mboxcl/mboxcl2 where some emails happen to have
Content-Length headers, despite being in an mboxo/mboxrd format.

Fix this by always requiring the format to parse, and return an error if
there is something suspicious going on, instead of trying to recover by
parsing into another format.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>