listing: don't forward events to self.component twice
Add a boolean variable flag to check whether we have forwarded an event
to self.component already before attempting to do it again. There
already was a check for it, but it did not account for non-consumable
read-only events such as AccountStatusChange.

This had the effect of account status change messages from the backend
such as "Establishing TLS connection" appearing twice while an IMAP
account was loading.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>