listing: move cmd_buf logic to Context
The `cmd_buf` in `Listing` component holds integer values, and when a
movement is performed, such as "move cursor down" or "page up" if it
holds a non-zero value, it is multiplied by that amount. So, pressing
"2" and then "PageDown" would move two pages forward.

Since we'd like to re-use this logic on any component that implements
scrolling or other integer command behavior, such as selecting an
attachment by index, let's move this to `Context`.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>