terminal: disable wraparound mode
It appears that on specific conditions we "overflow" when printing cells
on a row and the wraparound mode being enabled in the terminal causes
the cursor to go the start of the next line, causing a whole bunch of
rendering bugs.

I think the condition for this to happen is when there's a mismatch
between the width of a grapheme meli calculates and the actual amount of
cells it takes when printed to the terminal. In my setup, meli on a
terminal works without bugs, but when combined with tmux
(terminal->tmux->meli) the bug can appear when displaying wide
characters, usually on the status line.

In any case, we never want to wrap around, ever, so disable "Wraparound
Mode (DECAWM)" or `?7`.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>