Add tab_width option for terminal drawing
How we handled tab characters was inconsistent. We generally draw them
as two spaces, but for embedded grid output (used for filtering email
with external commands) the tab character was not drawn because wcwidth
returned zero width.

Use a user-provided tab width value with default value = 4 for drawing
tabs anywhere.

Signed-off-by: Manos Pitsidianakis <manos@pitsidianak.is>